:recycle:重构了上传图片的通用方法(可检验敏感、可存储图片)

+++ cloud/checkContent/checkImage/index.js
@@ -14,32 +14,48 @@ async function checkImage(data, cloud) {
 			fileID: data.value,
 		})
 		buffer = res.fileContent
-		// 异步删除验证用图片
-		cloud.deleteFile({
-			fileList: [data.value],
-		})
 	} catch (err) {
 		console.error('readFileSync Error.', err)
 	}
-
 	// console.log('downloadFile3', JSON.stringify(buffer))
 
 	if (buffer) {
 		try {
 			const res = await cloud.openapi.security.imgSecCheck({
 				media: {
-					// header: { 'Content-Type': 'application/octet-stream' },
 					contentType: 'image/png',
 					value: buffer,
 				},
 			})
+
 			console.log('checkImage', res)
+			if (res && res.errCode === 0) {
+				// 非敏感图片,如果路径是temp则删除
+				const nEnd = data.value.lastIndexOf('/')
+				const nStart = nEnd - 4
+				const str = data.value.substring(nStart, nEnd)
+
+				if (str === 'temp') {
+					cloud.deleteFile({
+						fileList: [data.value],
+					})
+				}
+			} else {
+				// 敏感图片一定删除
+				cloud.deleteFile({
+					fileList: [data.value],
+				})
+			}
 			objResult = {
 				code: 200,
 				data: res,
 			}
 		} catch (err) {
 			console.error('checkImage', err)
+			// 敏感图片一定删除
+			cloud.deleteFile({
+				fileList: [data.value],
+			})
 		}
 	}
